Use int instead of long to keep CSV row number
Generate TooManyEntriesNotice (error) for files that have more than
MAX_INT rows. Such large files would require too much memory and cause
OOM.

+++ core/src/main/java/org/mobilitydata/gtfsvalidator/notice/InvalidFloatNotice.java
@@ -23,12 +23,12 @@ package org.mobilitydata.gtfsvalidator.notice;
  */
 public class InvalidFloatNotice extends ValidationNotice {
   private final String filename;
-  private final long csvRowNumber;
+  private final int csvRowNumber;
   private final String fieldName;
   private final String fieldValue;
 
   public InvalidFloatNotice(
-      String filename, long csvRowNumber, String fieldName, String fieldValue) {
+      String filename, int csvRowNumber, String fieldName, String fieldValue) {
     super(SeverityLevel.ERROR);
     this.filename = filename;
     this.csvRowNumber = csvRowNumber;
